# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SD Payment API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authentication Method for Token Generation",
                "description": "This method facilitates user authentication and returns a token, which should be used for accessing subsequent methods that require bearer token authentication. It ensures secure interaction by validating user credentials and providing a token, essential for authorized access to restricted functionalities within the system.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "grupo1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "This method allows ending the session",
                "description": "The method is accessible with authentication and takes a username as a parameter for additional validation.",
                "operationId": "logout",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "grupo1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
        "/api/payments": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Generate payment",
                "description": "Generate payment.",
                "operationId": "generatePayment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": "100.00"
                                    },
                                    "information": {
                                        "type": "string",
                                        "example": "Payment for services"
                                    },
                                    "expirationDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "referenceDetails": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "externalId": {
                                                    "type": "integer",
                                                    "example": "280209"
                                                },
                                                "entity": {
                                                    "type": "integer",
                                                    "example": "28597"
                                                },
                                                "reference": {
                                                    "type": "integer",
                                                    "example": "049959124"
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": "100"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/payments/{paymentId}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get payment details",
                "description": "Get payment details.",
                "operationId": "getPaymentDetails",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "ID of the payment",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "externalId": {
                                            "type": "integer",
                                            "example": "280209"
                                        },
                                        "entity": {
                                            "type": "integer",
                                            "example": "28597"
                                        },
                                        "reference": {
                                            "type": "integer",
                                            "example": "049959124"
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": "100"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
        "/api/testconnection": {
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "A method to ensure services are reachable by their names and ports in a distributed system.",
                "description": "The primary purpose of this method is to verify if a specific service is accessible through its assigned name and associated port.",
                "operationId": "testconnection",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "endpoint": {
                                        "type": "string",
                                        "example": "http://webapp-service:3000/events"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get contents"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}